LayoutTests/http/tests/local/fileapi/upload-zip-bundle-as-blob-expected.txt

Test that bundles are automatically zipped when uploaded with XMLHttpRequest. To test manually, please drag a file bundle to a "Choose File" input below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS String.fromCharCode(responseArray[0], responseArray[1]) is 'PK'
PASS successfullyParsed is true

TEST COMPLETE
